Added support for setting table ID when it is larger than 0xff.
For the record, the semantic for setting the table ID is:

If the table ID is <= 0xff, then we set it in rtmsg->rtm_table,
otherwise we set rtmsg->rtm_table to RT_TABLE_UNSPEC and add the
real value as an RTA_TABLE attribute.

Note that table ID support for values large than 0xff has been
added to Linux around kernel version 2.6.20 or so.

Requested by:	Ben Greear <greearb AT candelatech.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10191 26435e26-4b25-483a-942d-477236992e84