Refactor the mechanism for pushing interface information so now it works with VLANs.
As part of the refactoring, now the IfConfigSet plugin API is
more generic and works better.

Note that future refactoring should take into consideration the
following:
Some IfTree state is setup from the top of the FEA (i.e., by user
configuration), other state is setup from the bottom of the FEA
(i.e., the system's kernel), while some state might be set by
either the top or the bottom of the FEA (e.g., interface MTU).

Some of the remaining complexity comes from the fact that the
various interface/vif/address state is not separated based
on the above observation.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10341 26435e26-4b25-483a-942d-477236992e84