In the original design the OSPF part of the configuration could contain interface/vif/address state that did not need to exist in the system. Unfortunately a dependancy on the actually configured interfaces crept in over time.
It is now possible to configure any interface/vif/address state in the
OSPF configuration and it does not need to exist.

The extraction of the prefix length, mtu, interface ID and
link-local address are now only done when the interface/vif exists in
the interfaces block and is configured up.

It is no longer a fatal error for OSPFv3 to not have a link-local
address configured, although a warning message is generated to aid in
debugging.

The Link-LSA for an interface/vif is now only generated when the
interface/vif is configured up.

Bug found by:	Ben Greear <greearb AT candelatech.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10363 26435e26-4b25-483a-942d-477236992e84