The first time the neighbour state goes 2-Way or higher schedule a NeighborChange event, which will cause a DR election.
If the router priority of a neighbour changes then schedule a
NeighborChange event, previously in state Waiting the event would not
have been scheduled.

No longer stop processing when in state Waiting.

Previously any changes in the DR or BDR settings of a neighbours hello
packets would have scheduled a NeighborChange event. Now only changes
by the DR or BDR will scheduled a NeighborChange event.

If the router ID of a neighbour changed print an info message.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10395 26435e26-4b25-483a-942d-477236992e84