Fix the mechanism for joining L2 MAC addresses:
* We need to use a regular socket "socket(AF_INET, SOCK_DGRAM, 0)"
  for the ioctl(SIOCADDMULTI) and ioctl(SIOCDELMULTI) system calls
  instead of the pcap(3) derived file descriptor.

* On FreeBSD and DragonFlyBSD the MAC address needs to be setup
  inside sockaddr_dl aligned address storage in a special way
  that is different from NetBSD and OpenBSD (and Linux).

Note that as of FreeBSD-6.2, the FreeBSD kernel has some bugs
when it comes to using ioctl(SIOCADDMULTI) and ioctl(SIOCDELMULTI)
to manipulate directly the L2 multicast membership on interface.
In particular, the kernel doesn't deal with duplicate joins, the
leave requests fail, and "netstat -ian" fails to print the joined
groups. See the following email thread for details:

http://lists.freebsd.org/pipermail/freebsd-net/2007-February/013120.html

However, those problems have been fixed (by Bruce M. Simpson) in the
FreeBSD CVS repository. The RELENG_6 fixes are:

src/usr.bin/netstat/mcast.c,v 1.2.8.1 2007/03/14 14:10:22 bms
src/sys/net/if.c,v 1.234.2.18 2007/03/08 13:15:27 bms

After the fix, "netstat -gn" will print the L2 joined groups as well.
Those fixes are in FreeBSD-current as well so they should appear
in the next FreeBSD release.

It appears that DragonFlyBSD (as of DragonFlyBSD-1.8) has similar
problems, but there is no indication it has been fixed even in the
most recent DragonFly-1.10.1 release.

For the record, the ioctl(SIOCADDMULTI) and ioctl(SIOCDELMULTI)
mechanism seems to work fine on NetBSD-3.1 and OpenBSD-4.1.
However, on NetBSD the ifmcstat(8) program needs to be used
to print the L2 membership ("netstat -ian" or "netstat -gn" don't
print it).
It appears that OpenBSD doesn't have ifmcstat(8), and "netstat"
doesn't print that info, so it is unclear what OS tool can be used
for that purpose.
Apparently, the NetBSD source code for ifmcstat(8) from NetBSD can
be compiled as-is on OpenBSD and it actually works.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10401 26435e26-4b25-483a-942d-477236992e84