Critical IPv6-related bug fix for setting the sockaddr len field of the routes added to the kernel.
Previously it was explicitly (re)set to the rounded-up length
of the corresponding sockaddr structure. This had no effect for IPv4,
because the rounded-up value was same as the original value,
but created bogus entries for IPv6.
Note that if the IPvX::copy_out() method is used, there is no need
to explicitly set the sockaddr len field, because IPvX::copy_out()
will take care of it.

Bug found by:	Hansi <hantongs AT gmail.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10434 26435e26-4b25-483a-942d-477236992e84