Moved the mechanism for testing whether the data plane supports IPv4/IPv6 from the FibConfigForwarding plugin to a new plugin named IfConfigProperty (managed by IfConfig).
The change is triggered by the requirement that the test for
IPv4/IPv6 data plane property should be before anything else.

Strictly speaking the new plugin should be managed by a new
manager type (e.g., DataPlanePropertyManager with implementations
inside a new directory fea/data_plane/property/) instead of IfConfig,
because the data plane properties are not bounded to the
interface management mechanism.
However, given that currently it is not clear what the future
requirements for testing new properties will be (the only data plane
properties we test for are IPv4/IPv6) it is a bit premature to start
adding the extra complexity associated with such new manager.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10514 26435e26-4b25-483a-942d-477236992e84