When a LSA is removed from the database the timer associated with it should be stopped.
In the past whenever a LSA was removed from the database it would also
be invalidated which would cause the timer to be stopped. It is now
possible to remove an LSA from the database and not have to invalidate
it, which means that the timer may not be stopped. It is possible for
a LSA to be removed from the database and the timer to still be
running, when the timer fires a sanity check in the timer code
verifies that the LSA is still in the database (at the same index). It
has been reported that this verification is failing.

The new code now unconditionally calls invalidate which unconditionally stops
the timer, which fixes an obvious problem.

I have not been able to reproduce problem.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=737
Bug found by:	Ben Greear <greearb AT candelatech.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10595 26435e26-4b25-483a-942d-477236992e84