Rate limit the transmission of STCP keep-alives.
There is a clearly reproducible problem in the XRL layer during times of
high system load. Whilst I have not examined the problem in detail,
the symptom and the fix are reproducible and consistent.

This change is an adaptation of the one originally attached to the
Bugzilla entry. With the change to use CLOCK_MONOTONIC in the
SystemClock class, this fixes the bug, as all timestamps used by
the XORP scheduler are now relative to a central clock source
independent of the wall-clock time.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=157
Bug found by:	Michael Larson


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10615 26435e26-4b25-483a-942d-477236992e84