Use memset() to initialize part of junk_seed.
This is needed to pacify a warning generated by the
gcc-4.1.2 compiler on Gentoo 2008.0_beta1 and using "jam"
for the compilation.
It appears that the warning was generated only if the
"-O" and "-g" compiler flags were used.
Apparently, the "-0" flag is added by "jam", but is
not added when using the default "gmake" compilation.

Bug found by:	Bruce M. Simpson


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10635 26435e26-4b25-483a-942d-477236992e84