Add a regression test for link layer access.
The test is enabled only for FreeBSD and Linux hosts. In FreeBSD,
the edsc(4) driver is used, and the test must be run as root
to instantiate an instance as it's a cloneable interface. In Linux,
the lo(4) driver may be used for link-layer loopback.

Regen bootstrap.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=508


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10649 26435e26-4b25-483a-942d-477236992e84