Fix incorrect STL container usage.
BufferedAsyncReader::set_reserve_bytes() is a method which may cause
its internal buffer, implemented as a std::vector, to be resized. Pointers
to elements of a std::vector are NOT guaranteed to remain valid after
the resize() method is called.

std::vector::resize() guarantees that elements are copied even if the
storage is reallocated by the default allocator. Therefore, cache the
_config.head pointer as an index before calling it, and restore after.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=750


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10667 26435e26-4b25-483a-942d-477236992e84