Bug fix in the Dummy ifconfig implementation.
Now the IfConfigSetDummy::push_config() pushed IfTree with the
configuration is saved locally inside the IfConfigSetDummy plugin
and the IfConfigGetDummy plugin returns it when
IfConfigGetDummy::pull_config() is called.

Previously the pushed IfTree state was saved in the top-level
IfConfig::system_config IfTree, which itself
was resetted when IfConfig::pull_config() was called.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10683 26435e26-4b25-483a-942d-477236992e84