Fix bugzilla #567.  I don't know if it's completely fixed, but it definitely is for connected routes.  Static routes behaves properly too.
The problem was route pushing.  Filters need to be rerun on ORIGINAL routes, not
the "last route", which could have been modified by filters.

Route pushing was never designed but just coded in a rather brute-force way in
protcols.  It should deserve attention in the future and should be implemented
by the policy framework.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10855 26435e26-4b25-483a-942d-477236992e84