Provide enhancement posted in Bugzilla #674.  You can now have length modifiers in network lists.
I was debating whether to make use of XORP's front-end parser operators or not.
This would catch syntax errors in the front-end.  For example, one solution
would have been something like:

modifier <= ""

That is, use the keyword modifier, and the actual modifier is the operator, and
then supply a dummy empty string or something.

For now I use the string itself as the operator, e.g.:

modifier: "<="


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10856 26435e26-4b25-483a-942d-477236992e84