allow for the "final action" in an unnamed term.  I had to write a little patch to get the rtrmgr order right.  I don't know whether it's an issue with the rtrmgr's ordering, and if so it should be fixed there - not in my policy patch. I don't know enough about the rtrmgr and all that ordering stuff looks kind of awkward, at least in use, in policy statements.
The juniper manual is vague on what the final action really is.  It seems to be
only accept/reject and those sorts of actions (maybe flow control too such as
next policy).  It doesn't seem to include route modification actions.  Also in
order documentation it hinted that terms that match, but with no action, take
the final action - the latest documentation omits this though.  (We don't take
the final action on terms that match but have no action.)

In practice, i think that the final action is used merely as a final
catch-allaccept/reject statement and we do provide this functionality.  From the
juniper manual it's unclear to me whether it is anything more than this.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10872 26435e26-4b25-483a-942d-477236992e84