Bug fix when checking whether the system has <linux/mroute.h>
The Linux multicast-related header files have various issues,
so we always need to use some hacks and work-arounds to get
them working.
In this case we need to use a hack to exclude <linux/in.h> that
might be included by <linux/mroute.h>, because <linux/in.h> will
conflict with <netinet/in.h> that was included earlier.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10913 26435e26-4b25-483a-942d-477236992e84