Force commit to correct CVS log message.  Message /tmp/log.r486 should read:
Fix a small "bug".  Run timers that have expired exactly now - the rest of the
code assumes this will happen, but it actually doesn't.  In practice it means
things are a bit slower because we need to execute the eventloop twice each time
a timer expires exactly.  (This will be important for a future commit i do.)


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10974 26435e26-4b25-483a-942d-477236992e84