Force commit to correct CVS log message.  Message /tmp/log.r492 should read:
All tests pass with new XRL code.  The only hack was adding a 1 second sleep to
olsr's test_routing1.py test_c4_partial test.  I need to understand this better
but I think that it has to do with the simulator trying to keep events
synchronized with time and assuming that each eventloop run is a "single
iteration" - this no longer holds with the aggressive eventloop.  We need to
decide whether we want this new eventloop (+26% performance with XRL benchmark)
or whether we should keep the old one for simplicity and safety.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10978 26435e26-4b25-483a-942d-477236992e84