The test_peering2.sh tests have recently been failing regularly, and in the past have failed intermittently.
The coord process can accept one command at a time. A pending command
is provided to make sure that the previous command has completed. The
shell function that sends a command calls pending once a second for
five seconds waiting for the command to complete. If the command does
not complete an error is generated and the test fails.

It looks like the five second delay is not long enough for some
commands to complete. The refactoring of the eventloop has made some
tests fail every time. The delay has been increased to ten seconds and
a warning placed in the code to make it obvious that the command has
not really failed, just timed out.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10998 26435e26-4b25-483a-942d-477236992e84