When requested to establish a BGP session carry out the full handshake, including sending the keepalive when the peer that initiated the session receives an open in response to its open.
Previously the peer would only generate a keepalive in response to a
peers keepalive, which meant that the time to establish a session
would be determined by the holdtime of the BGP under test.

The holdtime of the BGP tests in test_peering2.sh was set to five
seconds, the time that the tests waited for a session to be
established was five seconds. This was the actual cause of the
intermittent failures seen in the tinderbox runs.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10999 26435e26-4b25-483a-942d-477236992e84