Implement proxy ARP within XORP.  For now, only VRRP uses it, but it is simple to factor out "arpd" and make it available to the whole router.  We could use OS proxy arp mechanisms but that's probably a nightmare to figure out right now - it was easier to code a built-in mechanism.
This is used by VRRP when a backup router becomes a master, and the backup
router doesn't actually own the IP addresses of the virtual router.  When this
occurs, we still need to reply to ARP requests for those addresses.

(Note our VRRP implementation doesn't handle the "opposite" case of a backup
router owning some of the IP addresses - it will still send ARP replies for
those addresses when it shouldn't.  This will probably remain unimplemented
until someone shouts.)


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11054 26435e26-4b25-483a-942d-477236992e84