A problem was introduced in the performance enhancements: http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=808
A pointer was held in an Xrl to a FinderDBEntry the entry contains the
resolved Xrl. The FinderDBEntry is held in a map so it is not valid to
hold a pointer to it, as any changes to the map could cause the
invalidate the pointer. Another problem was that if the target of the
Xrl died then the cached FinderDBEntry would not be uncached.

A solution to the problem may be to hold the FinderDBEntry as a
pointer in the map so that updates would not move the pointer. Another
solution is to hold a pointer to the resolved Xrl in the unresolved
Xrl and not have the intermediate FinderDBEntry. The FinderDBEntry
holds a list of resolved Xrls, it is not clear why it is a list and
not a single entry as they all resolve the same target, unless it is a
list to hold multiple request responses, in which case the list
shouldn't be exposed outside the FinderClient.

The code has been reverted to not caching the FinderDBEntry but the
XrlPFSender is now being cached in the Xrl and resolved is now a bool
not a FinderDBEntry. There will be a slight drop in performance doing
the FinderDBEntry each time but this is not the correct pointer to
hold. The FinderClient needs to be reworked to return an Xrl pointer
that can be safely stored.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=800


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11153 26435e26-4b25-483a-942d-477236992e84