As noted by Bruce and Pavlin this test has been failing for two reasons one the burst size of 50 is too high and secondly the scheduling code in drip_run did not correctly cater for slow systems.
The failure in the test was not a problem with the finder subsystem
but a problem with the tests.

The tests are now totally event driven, previously they were a mix of
event and timer driven and on a slow system the timer code broke the
tests.

The burst size has been reduced to 25.

Now the tests are totally event driven to protect against a missing event a
watchdog of 1 minute is started to guard against missing event.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=799


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11156 26435e26-4b25-483a-942d-477236992e84