Fix bug 793.
This actually fixes another bug which indirectly fixes 793.  When a process is
born, we reconfigure its filters and then push routes.  Before, the logic would
not push routes to avoid extra work (pushing is expensive) because the project
just started so it has no routes.  Well, there's no guarantee that the process
doesn't aquire routes by the time the filters are actually configured.  We
therefore have to push routes anyway.  Processes that want to avoid this extra
work can wait, while booting up, for this push message which implicitly
indicates that filter configuration has finished.  BGP does this for example.

Bug 793 existed becuase BGP would wait for this message that was never sent due
to the policy manager "optimization" (aka. avoid push bug).  In practice this
bug is difficult to trigger because BGP always configures its filters (even with
a "null" configuration) which will causse a route push.  However if the policy
manager received BGP's configuration prior to the notification of BGP's birth,
the bug is triggered.  Typically the events occur in reverse order (hence not
causing the bug), but this is purely based on scheduling.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=793


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11157 26435e26-4b25-483a-942d-477236992e84