Import the Windows development document. Not yet connected to the build.
This is a technically rigorous and (hopefully)
fairly complete engineering document, which describes
in detail how XORP has been ported to the Windows platform.

Note: The UML figures are in SVG format. This is not yet
supported natively in LaTeX, conversion will be needed,
so they are stubbed out.

Requested by:	Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11178 26435e26-4b25-483a-942d-477236992e84