Add conditional inclusion of graphicx package with dvipdf. Use an explicit scale and bounding box in the includegraphics command.
The diagram is still present in the PDFLaTeX output, and both
PDFLaTeX and LaTeX can now build the file.
However, page 5 is currently blank in the PostScript output; the
diagram does not appear when the DVI is viewed manually.

The workaround to achieve PostScript output is to export
PostScript from the PDF generated with PDFLaTeX. The real solution
for this issue is to wait for native SVG support in LaTeX.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11180 26435e26-4b25-483a-942d-477236992e84