Fix an issue with OSPF virtual links, where packets destined for a virtual link endpoint would contain a TTL of 1, resulting in them being dropped as they transited the network.
The downcall to the FEA to transmit a packet destined for a virtual
link endpoint would rely on the FEA to set the TTL of the outer IPv4
header.

As the OSPF process passes the ip_network_control flag true to the
FEA when it sends the tranmission XRL, and a ttl of -1 to let the FEA
choose the IPv4 TTL, this resulted in the FEA setting the TTL to 1
internally every time.

The fix is to explicitly set the TTL for tranmission from within the
OSPF PeerManager::transmit() method, as it has knowledge of whether or
not the endpoint is virtual.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=824
Reviewed by:	Pavlin Radoslavov


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11188 26435e26-4b25-483a-942d-477236992e84