If the designated router crashes then its Network-LSA will be sent back to it when it restarts, the protocol requires that this old Network-LSA is responded to by a MaxAge LSA.
If OSPF received a MaxAge LSA while any of its neighbours were in state
EXCHANGE or LOADING then OSPF would send the MaxAge LSA straight
back, in other states the MaxAge LSA would be delay acked and the LSA
would be dropped.

The behaviours described above meant that when an adjacency was being
formed a received MaxAge LSA would immediately be responded to by a
MaxAge LSA, causing a continious flow of update packets.

When in state EXCHANGE or LOADING an incoming MaxAge LSA is now
ignored and the add_lsa has an ASSERT to make sure that a MaxAge LSA
cannot be stored in the LSA database.

Bugzilla URL:	http://bugzilla.xorp.org/bugzilla/show_bug.cgi?id=785


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11236 26435e26-4b25-483a-942d-477236992e84