Changed make rules for clnt-gen and tgt-gen to reflect the fact that the generated source and header files are written only to $(srcdir), and not both $(srcdir) and the build directory.
This avoids problems when *.xif files are compiled.  Currently, when
gmake invokes *-gen on a *.xif file, the generated *.cc and *.hh files
are stored in $(srcdir).  But because gmake thinks the files were also
stored in the build directory, it uses that path when it subsequently
invokes the compiler.  Since the files don't exist there, the build
will fail.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11348 26435e26-4b25-483a-942d-477236992e84