Changed semantics of IPvX constructors, copy_in, and copy_out methods that operate on sockaddr* structs.
Previously, these would work if a sockaddr struct of a different type
was passed as an argument.  For example, a sockaddr_in6 could be cast
to a sockaddr_in and passed to the cooresponding ctor or copy_in
method, and it would be loaded as an IPv6 address without error.
Similarly, a sockaddr_in could be cast to a sockaddr_in6 and passed to
the cooresponding ctor or copy_in method, and it too would be loaded
as a IPv4 address without error.  The same was true with copy_out,
even though copying out a IPv6 address into a sockaddr_in struct could
result in memory overflow.

In bugzilla issue 1076, a problem was reported that these casts from
sockaddr_in to/from sockaddr_in6 was causing a "cast from 'sockaddr_in*'
to 'sockaddr_in6*' increases required alignment of target type" warning/
error.

I have changed these functions so that passing a sockaddr of the wrong
"family" now causes a InvalidFamily exception to be thrown, as was
already for other address families (AF_UNSPEC, AF_LOCAL, etc.).

However, this change may not be sufficent to get XORP to run on strict
alignment machines.  In libxorp/utility.h, functions like
sockaddr2sockaddr_in perform similar casts as the ones that I removed,
but do so with an intermediate cast to a void pointer.  This avoids
the gcc warning/error, potentially leaving the alignment problem to
run time.

I also installed an ELDK ARM cross development environment and found
additional alignment problems elsewhere in the XORP code.

In libproto/packet.hh, a uint8_t ptr is cast to an ArpHeader, and in
vrrp/vrrp_packet, a uint8_t ptr is cast to a VrrpHeader.  It appears
additional work is required for VRRP (even the ARP code is only used
by VRRP) to work on strict alignment machines.

Bugzilla URL:	http://bugzilla.xorp.org/show_bug.cgi?id=1076


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11375 26435e26-4b25-483a-942d-477236992e84