Changed XRL target generator to emit a const, static array containing the name and the callback method for each operation.
This table is used with a loop when adding and removing handlers.  The
old implementation open-coded a sequence of calls to add_handler() and
remove_handler().

The new implementation results in about a 30% reduction in code size.
This is larger that might be expected because the add_handler() and
remove_handler() methods take C++ string arguments, so string objects
needed to be constructed from string literals before each call, and
destructed afterwards.  We may be able to achieve a small amount of
additional code size reduction by changing those methods to take a
const char ptr, which would avoid the construction and destruction of
the string.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11377 26435e26-4b25-483a-942d-477236992e84