When reconfiguring bootstrap router zones in XORP's PIM implementation, calling stop() followed by start() will blow away the Elected-BSR state if the XORP router is the elected BSR for a given multicast zone.
Add an explicit apply_bsr_changes method to PIM's XRL RPC interface, but
do not instruct the Router Manager to use it yet, as this change still
requires more rigorous testing in the field.

Users are encouraged to test this change by uncommenting the %update
change in PIM's template file which will be committed in a forthcoming
revision.

Submitted by:	Samuel Lucas Vaz de Mello, Pavlin Radoslavov (with fixups)


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11529 26435e26-4b25-483a-942d-477236992e84