Correct the minimum length of an AS-External-LSA in OSPFv3.
As per RFC 2740 section A.4.1, if the PrefixLength is 0, i.e. a default
route is being advertised, the Address Prefix itself may be completely
elided from the on-wire encoding. An illustration of this may be found
in section A.4.7.

Bugzilla URL:	http://bugzilla.xorp.net/show_bug.cgi?id=1059
Submitted by:	Petr Lampa


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11539 26435e26-4b25-483a-942d-477236992e84