In Neighborhood::consider_remaining_cand_mprs(), do not screen out N already designated as MPRs when populating the cand-MPR set; this is premature optimization.
Candidate MPR selection needs to consider the whole MPR space, even
if existing candidates are likely to win. Just because an N was
previously chosen as an MPR to cover a poorly reachable N2, does not
mean all its N2 are covered by the time this method runs.

Submitted by:	Ben Greear




git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11562 26435e26-4b25-483a-942d-477236992e84