When a callback is added whilst a callback is currently being serviced, std::vector::resize() will be called, which invalidates 'this' for the SelectorList::Node currently being accessed. Freed heap memory is accessed, further use of this could result in C++ heap corruption.
Preallocate 256 file descriptor slots (Node) in the constructor for
the std::vector SelectorList::_selector_entries. 256 is chosen to be
conservative (most processes won't be using this many).

This is not the right fix, this is purely a temporary workaround for
the problem, and it does waste memory:
 (((sizeof(int)) + (sizeof(int)) + (sizeof(int)) +
  ((sizeof(int) + sizeof(int) + sizeof(void *) + sizeof(void *))) * 3) * 256
 ==> 21504 bytes on the IA32 architecture, for each SelectorList instance.

The debugging portion of the patch has been omitted from this commit,
as a partial rewrite is probably called for (although the type initializer
stuff is a really good idea).

Submitted by:   Ben Greear
Found with:     valgrind


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11566 26435e26-4b25-483a-942d-477236992e84