Removed "arch", "os", and "cross" command line variables, and replaced them with "host" and "build". These variables take a system tripple or system alias, just like the --host and --build command line options of a GNU configure script.  The host and build system tripples are broken down further into cpu, vendor, and OS; for use throughout the build and configuration infrastructure.
Remove "rel" command line variable, as it was only used as a suffix for
the build directory.

While I could have written our own SCons/python code to guess the
build system and to canonicalize system tripples; instead I've added
the most current versions of GNU config.guess and config.sub scripts
and have the SConstruct call them.  This ensures the results will be
the same as other packages that use the GNU Build System.  This may
be something we want to revisit.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11572 26435e26-4b25-483a-942d-477236992e84