Add support for running regression tests in a shared library world. With this change, 'scons check shared=true' runs to completion.
Note: ORIGIN support is not yet enabled. The shared library paths
are hardcoded to the $BUILDDIR. This hard-coding happens in the ELF
.dynamic section (specifically the DT_RPATH entry) of each compiled test
program.

The rtld RPATH feature needs an expanded (i.e. non-SCons) path name,
as the SCons builders pass it directly to the linker.
This path changes if we install all .so into a $libdir (new XORP layout),
or if we install them into a subdir of the destination tree for each
subsystem (current XORP layout).

For tests, we want RPATH directories to be passed either as absolute
paths rooted at $BUILDDIR, or relative to the executable (../../subdir)
if using ORIGIN.

If we use rtld's ORIGIN feature to let it find our .so using
a path relative to where the Program is installed, we need
to turn each $BUILDDIR in $RPATH into a path relative to the Program.
This requires SCons is run under Python 2.6 or later, as os.path.relpath()
is used to determine the relative path for ORIGIN.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11587 26435e26-4b25-483a-942d-477236992e84