Bite the bullet, and build with shared libraries by default.
This is a reasonable assumption, given that we now have regression
test support for a shared library build, and the tests pass.
The commercial product is also using shared libraries by default.

Not all of the tree will fully use shared libraries. BGP in particular
needs to be revisited, because of some inter-module coupling.

Static libraries will continue to be semi-supported for debugging purposes,
however, it's expected that they will *not* co-exist in the same build run.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11592 26435e26-4b25-483a-942d-477236992e84