Add a 'debug' command line variable which can be used to set the debug symbol level in a portable way. CFLAGS/CXXFLAGS will be sanitized to enforce this option, unless 'debug' is set to 'override'.
This switch only affects the build as it is being staged; if 'strip'
is enabled, debug information will be stripped on a 'scons install'
as before.

Set the default to 'full', which passes -g3 to gcc and g++. This seems
reasonable, as most folk building from SVN will want full debug information.
On ELF platforms, DWARF-2 normally will be used, and -g3 will produce
richer C++ debugging information in the DWARF sections.

Update comments.

Approximate disk usage of $BUILDDIR for an amd64 self-hosted build,
with optimize=yes (gcc/g++ -O1):-
 705M with debug=yes (the old default)
 831M with debug=full (the new default)


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11596 26435e26-4b25-483a-942d-477236992e84