Fix memory leak and non-reentrant behaviour in the XRL client stubs. The fix is applied to the code generator.
Previously, individual Xrls for each client method were cached
using a static pointer to the class Xrl instance. This meant that the
allocated memory could not be reclaimed, and the code was non-reentrant.

Allocation is now made through an auto_ptr<Xrl*> private to the
client stub class. If the client stub class is re-instantiated repeatedly,
this may cause allocator churn.
In practice, most XORP code instantiates a client only when it needs
to make a call; this may need to be revisited.

'scons check' runs without error on FreeBSD 7.2-STABLE/amd64.

With: find obj/*/xrl/interfaces -name '*.so' | xargs size -t
+54K code size on amd64 with optimize=no.
+41K code size on amd64 with optimize=yes.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11597 26435e26-4b25-483a-942d-477236992e84