Commit top level support for rtld ORIGIN as an experimental build feature. Default to disabled.
ORIGIN support allows relative paths to shared libraries
to be embedded in binaries, instead of absolute ones. XORP uses
a lot of shared libraries of its own, so may benefit from this support.

Unfortunately, RPATH ORIGIN support hasn't been MFC'ed yet to FreeBSD
7.2, so I cannot fully test this change just now.

Also, because we currently install binaries in various weird and
wonderful places which roughly match to the layout of the source tree,
RPATH would need to be set individually for each set of binaries with
a system location.

The use of os.path.relpath() also requires Python 2.6 to run the
SConstruct when this feature is in use.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11626 26435e26-4b25-483a-942d-477236992e84