Expand the use of $ORIGIN in the runtime ELF linker. Do not enable it by default yet.
The trick here is to always use a relative RPATH of $ORIGIN/../lib,
and to mimic the installed layout in $BUILDDIR using symlinks.

NOTE WELL: This change is not complete; some binaries are installed
further down in the target tree, so using a globally relative path
will break. This change should, however, allow you to run most
binaries from within BUILDDIR without installing during development.

The conditional spamming in each SConscript can be toned down if we
override the behaviour of the existing SharedLibrary builder in
our SConsEnvironment to use our PostActions.
For now, we make the creation of the symlink explicit where needed.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11643 26435e26-4b25-483a-942d-477236992e84