Continue implementation of FHS style package layout for XORP.
With this change, binaries are now installed as follows:
$prefix/sbin                    xorpsh, router manager and profiler
$prefix/lib/xorp/bin            anything in */tools/*
$prefix/lib/xorp/sbin           all protocol modules
$prefix/lib/xorp/lib            all shared libraries
$prefix/share/xorp/templates    templates are arch independent data
$prefix/share/xorp/xrls         *.xrls if debug_xrldb is enabled

Fix some fallout from the last commit, where exec_prefix was
set as a local in SConstruct, but not added to the SCons environmnet.
Use explicit env['foo'] references, rather than relying on '$foo',
to catch this sort of thing sooner in the dependent SConscripts.

NOTE WELL: Trunk at this revision will NOT run, until I've
finished updating the Router Manager for the FHS style paths.
This should happen within the next 24 hours.
If you need a working checkout from trunk, use revision 11646.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11648 26435e26-4b25-483a-942d-477236992e84